refactor pr based on review suggestion
