+++ flink-yarn/src/main/java/org/apache/flink/yarn/entrypoint/YarnEntrypointUtils.java
@@ -98,10 +98,8 @@ public class YarnEntrypointUtils {
 			configuration.setInteger(WebOptions.PORT, 0);
 		}
 
-		if (!configuration.getString(RestOptions.BIND_PORT).equals("0")) {
-			// set the REST port to 0 to select it randomly
-			configuration.setString(RestOptions.BIND_PORT, "0");
-		}
+		// set the REST port to 0 to select it randomly
+		configuration.setString(RestOptions.BIND_PORT, "0");
 
 		// if the user has set the deprecated YARN-specific config keys, we add the
 		// corresponding generic config keys instead. that way, later code needs not
